src/main/java/attractions/RollerCoaster.java
package attractions;

import interfaces.ITicketed;
import themepark.Visitor;

public class RollerCoaster extends Attraction implements ITicketed {

    public RollerCoaster(String name, int rating) {
        super(name, rating);
    }

    @Override
    public double defaultPrice() {
        return 8.40;
    }

    @Override
    public double priceFor(Visitor visitor) {
        if (visitor.getHeight() > 2.00) {
            return defaultPrice() * 2;
        }
        return defaultPrice();
    }
}
